Deleted AuthorizationServlet, and added LoginServlet and LogoutServlet. Created first jsp file that named login.jsp for checking cookies of user and for exiting from the site. Created connection to the database in ContextListener.

+++ src/main/java/org/computer/school/schedule/app/servlet/entities/ServletURIProcessing.java
@@ -1,5 +1,6 @@
 package org.computer.school.schedule.app.servlet.entities;
 
+import org.computer.school.schedule.app.datastorage.db.DBConnection;
 import org.computer.school.schedule.app.servlet.WebPageMessage;
 import org.computer.school.schedule.app.servlet.entities.operations.ServletEntityOperations;
 import org.computer.school.schedule.app.servlet.entities.operations.CoursesServletOperations;
@@ -10,6 +11,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.sql.Connection;
 import java.util.Map;
 
 /**
@@ -19,19 +21,20 @@ public class ServletURIProcessing extends HttpServlet implements ServletURIProce
     private HttpServletRequest request;
     private HttpServletResponse response;
     private String[] pathTokens;
-
     private ServletEntityOperations servletOperation;
     private WebPageMessage m;
-
-    public ServletURIProcessing() {
-    }
+    private Connection databaseConnection;
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         this.request = request;
         this.response = response;
-
         m = new WebPageMessage(response);
+
+        /* Getting connection to the database */
+        DBConnection dbConnection = (DBConnection) getServletContext().getAttribute("dbConnection");
+        databaseConnection = dbConnection.getConnection();
+
         URIProcessing();
     }
 
@@ -43,44 +46,50 @@ public class ServletURIProcessing extends HttpServlet implements ServletURIProce
             pathTokens = pathInfo.split("/");
             if (pathTokens.length > 1) {
                 pathParams = request.getParameterMap();
-                switch(pathTokens[1]) {
-                    case "courses": {
-                        servletOperation =
-                                new CoursesServletOperations(
-                                        response, pathParams, pathTokens);
-                        break;
+                String entity = pathTokens[1];
+                if(!entity.isEmpty()) {
+                    switch (entity) {
+                        case "courses": {
+                            servletOperation =
+                                    new CoursesServletOperations(
+                                            databaseConnection, response, pathParams, pathTokens);
+                            break;
+                        }
+                        case "persons": {
+                            servletOperation =
+                                    new PersonsServletOperations(
+                                            databaseConnection, response, pathParams, pathTokens);
+                            break;
+                        }
+                        default:
+                            m.message("No such entity!");
                     }
-                    case "persons": {
-                        servletOperation =
-                                new PersonsServletOperations(
-                                        response, pathParams, pathTokens);
-                        break;
-                    }
-                    default:
-                        m.message("No such entity!");
+                    choiceServletOperation();
                 }
-                choiceServletOperation();
             }
         }
     }
 
     @Override
     public void choiceServletOperation() {
-        switch (pathTokens[2]) {
-            case "read": {
-                servletOperation.read();
-                break;
-            }
-            case "insert": {
-                servletOperation.insert();
-                break;
-            }
-            case "delete": {
-                servletOperation.delete();
-                break;
-            }
-            default: {
-                m.message("No such operation!");
+        String operation = pathTokens[2];
+        if(!operation.isEmpty()) {
+            switch (pathTokens[2]) {
+                case "read": {
+                    servletOperation.read();
+                    break;
+                }
+                case "insert": {
+                    servletOperation.insert();
+                    break;
+                }
+                case "delete": {
+                    servletOperation.delete();
+                    break;
+                }
+                default: {
+                    m.message("No such operation!");
+                }
             }
         }
     }
